+++ api/go.mod
@@ -7,7 +7,6 @@ require (
 	github.com/moby/docker-image-spec v1.3.1
 	github.com/opencontainers/go-digest v1.0.0
 	github.com/opencontainers/image-spec v1.1.1
-	golang.org/x/time v0.11.0
 	gotest.tools/v3 v3.5.2
 	pgregory.net/rapid v1.2.0
 )
