+++ daemon/builder/dockerfile/copy.go
@@ -16,10 +16,10 @@ import (
 	"github.com/containerd/log"
 	"github.com/moby/buildkit/frontend/dockerfile/instructions"
 	"github.com/moby/go-archive"
-	"github.com/moby/moby/api/pkg/progress"
 	"github.com/moby/moby/v2/daemon/builder"
 	"github.com/moby/moby/v2/daemon/builder/remotecontext"
 	"github.com/moby/moby/v2/daemon/builder/remotecontext/urlutil"
+	"github.com/moby/moby/v2/daemon/internal/progress"
 	"github.com/moby/moby/v2/daemon/internal/streamformatter"
 	"github.com/moby/moby/v2/daemon/internal/system"
 	"github.com/moby/moby/v2/pkg/longpath"
