+++ daemon/internal/distribution/xfer/download_test.go
@@ -12,8 +12,8 @@ import (
 	"time"
 
 	"github.com/docker/distribution"
-	"github.com/moby/moby/api/pkg/progress"
 	"github.com/moby/moby/v2/daemon/internal/layer"
+	"github.com/moby/moby/v2/daemon/internal/progress"
 	"github.com/opencontainers/go-digest"
 	"github.com/opencontainers/image-spec/identity"
 	"gotest.tools/v3/assert"
