+++ daemon/internal/distribution/xfer/transfer_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/moby/moby/api/pkg/progress"
+	"github.com/moby/moby/v2/daemon/internal/progress"
 )
 
 func TestTransfer(t *testing.T) {
