+++ daemon/internal/image/tarexport/load.go
@@ -17,11 +17,11 @@ import (
 	"github.com/docker/distribution"
 	"github.com/moby/go-archive/chrootarchive"
 	"github.com/moby/go-archive/compression"
-	"github.com/moby/moby/api/pkg/progress"
 	"github.com/moby/moby/api/types/events"
 	"github.com/moby/moby/v2/daemon/internal/image"
 	"github.com/moby/moby/v2/daemon/internal/ioutils"
 	"github.com/moby/moby/v2/daemon/internal/layer"
+	"github.com/moby/moby/v2/daemon/internal/progress"
 	"github.com/moby/moby/v2/daemon/internal/streamformatter"
 	"github.com/moby/moby/v2/daemon/internal/stringid"
 	"github.com/moby/sys/sequential"
