+++ daemon/internal/progress/progress.go
