+++ daemon/internal/progress/progressreader.go
