+++ daemon/pkg/plugin/fetch_linux.go
@@ -14,9 +14,9 @@ import (
 	"github.com/containerd/log"
 	"github.com/distribution/reference"
 	"github.com/moby/go-archive/chrootarchive"
-	"github.com/moby/moby/api/pkg/progress"
 	"github.com/moby/moby/api/types/registry"
 	progressutils "github.com/moby/moby/v2/daemon/internal/distribution/utils"
+	"github.com/moby/moby/v2/daemon/internal/progress"
 	"github.com/moby/moby/v2/daemon/internal/stringid"
 	"github.com/moby/moby/v2/pkg/ioutils"
 	"github.com/opencontainers/go-digest"
